[SPARK-29007][STREAMING][MLLIB] Enforce not leaking SparkContext in tests which creates new StreamingContext with new SparkContext
