+++ streaming/src/test/scala/org/apache/spark/streaming/BasicOperationsSuite.scala
@@ -28,6 +28,7 @@ import org.apache.spark.{HashPartitioner, SparkConf, SparkException}
 import org.apache.spark.rdd.{BlockRDD, RDD}
 import org.apache.spark.storage.StorageLevel
 import org.apache.spark.streaming.dstream.{DStream, WindowedDStream}
+import org.apache.spark.streaming.testutil.StreamingTestUtils._
 import org.apache.spark.util.{Clock, ManualClock}
 
 class BasicOperationsSuite extends TestSuiteBase {
