+++ streaming/src/test/scala/org/apache/spark/streaming/ReceiverInputDStreamSuite.scala
@@ -28,6 +28,7 @@ import org.apache.spark.streaming.dstream.ReceiverInputDStream
 import org.apache.spark.streaming.rdd.WriteAheadLogBackedBlockRDD
 import org.apache.spark.streaming.receiver.{BlockManagerBasedStoreResult, Receiver, WriteAheadLogBasedStoreResult}
 import org.apache.spark.streaming.scheduler.ReceivedBlockInfo
+import org.apache.spark.streaming.testutil.StreamingTestUtils._
 import org.apache.spark.streaming.util.{WriteAheadLogRecordHandle, WriteAheadLogUtils}
 
 class ReceiverInputDStreamSuite extends TestSuiteBase with BeforeAndAfterAll {
@@ -36,6 +37,7 @@ class ReceiverInputDStreamSuite extends TestSuiteBase with BeforeAndAfterAll {
     try {
       StreamingContext.getActive().foreach(_.stop())
     } finally {
+      ensureNoActiveSparkContext()
       super.afterAll()
     }
   }
