+++ streaming/src/test/scala/org/apache/spark/streaming/ReceiverSuite.scala
@@ -34,6 +34,7 @@ import org.apache.spark.storage.StorageLevel
 import org.apache.spark.storage.StreamBlockId
 import org.apache.spark.streaming.receiver._
 import org.apache.spark.streaming.receiver.WriteAheadLogBasedBlockHandler._
+import org.apache.spark.streaming.testutil.StreamingTestUtils._
 import org.apache.spark.util.Utils
 
 /** Testsuite for testing the network receiver behavior */
