+++ streaming/src/test/scala/org/apache/spark/streaming/TestSuiteBase.scala
@@ -34,6 +34,7 @@ import org.apache.spark.internal.Logging
 import org.apache.spark.rdd.RDD
 import org.apache.spark.streaming.dstream.{DStream, ForEachDStream, InputDStream}
 import org.apache.spark.streaming.scheduler._
+import org.apache.spark.streaming.testutil.StreamingTestUtils._
 import org.apache.spark.util.{ManualClock, Utils}
 
 /**
@@ -269,23 +270,6 @@ trait TestSuiteBase extends SparkFunSuite with BeforeAndAfter with Logging {
 
   before(beforeFunction)
   after(afterFunction)
-
-  /**
-   * Run a block of code with the given StreamingContext and automatically
-   * stop the context when the block completes or when an exception is thrown.
-   */
-  def withStreamingContext[R](ssc: StreamingContext)(block: StreamingContext => R): R = {
-    try {
-      block(ssc)
-    } finally {
-      try {
-        ssc.stop(stopSparkContext = true)
-      } catch {
-        case e: Exception =>
-          logError("Error stopping StreamingContext", e)
-      }
-    }
-  }
 
   /**
    * Run a block of code with the given TestServer and automatically
