+++ streaming/src/test/scala/org/apache/spark/streaming/ui/StreamingJobProgressListenerSuite.scala
@@ -25,6 +25,7 @@ import org.apache.spark.scheduler.SparkListenerJobStart
 import org.apache.spark.streaming._
 import org.apache.spark.streaming.dstream.DStream
 import org.apache.spark.streaming.scheduler._
+import org.apache.spark.streaming.testutil.StreamingTestUtils._
 
 class StreamingJobProgressListenerSuite extends TestSuiteBase with Matchers {
 
@@ -35,9 +36,7 @@ class StreamingJobProgressListenerSuite extends TestSuiteBase with Matchers {
 
   override def afterFunction() {
     super.afterFunction()
-    if (ssc != null) {
-      ssc.stop()
-    }
+    ensureNoActiveSparkContext(ssc)
   }
 
   private def createJobStart(
